Removed repeating value
