+++ src/utils/recordLikes.js
@@ -1,21 +1,7 @@
 import { ENV_API, APP_ID } from '../api/config.js';
 import showSnackbar from './showSnackBar.js';
 import fetchLikes from '../data/fetchLikes.js';
-
-export const updateLikesCount = (itemId, likes) => new Promise((resolve, reject) => {
-  const card = document.querySelector(`[data-item-id="${itemId}"]`);
-  if (card) {
-    const likeCount = card.querySelector('.like-count');
-    if (likeCount) {
-      likeCount.textContent = likes;
-      resolve(); // Resolve the promise when the likes count is updated
-    } else {
-      reject(new Error('Like count element not found'));
-    }
-  } else {
-    reject(new Error('Card element not found'));
-  }
-});
+import updateLikesCount from './updateLikesCount.js';
 
 export const recordLikes = (itemId) => {
   const url = `${ENV_API}${APP_ID}/likes/`;
@@ -42,3 +28,4 @@ export const recordLikes = (itemId) => {
       showSnackbar('Error recording likes!', error);
     });
 };
+export default recordLikes;
